refactoring and renaming ldap_htw_adapter
